+++ sparapy/__init__.py
@@ -10,14 +10,16 @@ __version__ = '0.1.0'
 from . import geometry
 from . import radiosity
 from . import sound_object
-from . import radiosity_fast
+from .radiosity_fast.radiosity_class import DRadiosityFast
 from . import testing
+from . import radiosity_fast
 
 
 __all__ = [
     'geometry',
     'radiosity',
     'sound_object',
+    'DRadiosityFast',
     'radiosity_fast',
     'testing',
 ]
