+++ sparapy/radiosity_fast/__init__.py
@@ -1,0 +1,11 @@
+"""Module of the fast radiosity solver."""
+from .radiosity_class import (
+    DRadiosityFast,
+)
+
+from . import geometry
+
+__all__ = [
+    'DRadiosityFast',
+    'geometry',
+]
