+++ sparapy/radiosity_fast/form_factor.py
@@ -1,0 +1,205 @@
+"""Form factor calculation for radiosity."""
+import numba
+import numpy as np
+from . import geometry
+
+
+@numba.njit(parallel=True)
+def kang(
+        patches_center:np.ndarray, patches_normal:np.ndarray,
+        patches_size:np.ndarray, visible_patches:np.ndarray) -> np.ndarray:
+    """Calculate the form factors between patches.
+
+    Parameters
+    ----------
+    patches_center : np.ndarray
+        center points of all patches of shape (n_patches, 3)
+    patches_normal : np.ndarray
+        normal vectors of all patches of shape (n_patches, 3)
+    patches_size : np.ndarray
+        size of all patches of shape (n_patches, 3)
+    visible_patches : np.ndarray
+        index list of all visible patches combinations (n_combinations, 2)
+
+    Returns
+    -------
+    form_factors : np.ndarray
+        form factors between all patches of shape (n_patches, n_patches)
+        note that just i_source < i_receiver are calculated ff[i, j] = ff[j, i]
+
+    """
+    n_patches = patches_center.shape[0]
+    form_factors = np.zeros((n_patches, n_patches))
+    for i in numba.prange(visible_patches.shape[0]):
+        i_source = int(visible_patches[i, 0])
+        i_receiver = int(visible_patches[i, 1])
+        source_center = patches_center[i_source]
+        source_normal = patches_normal[i_source]
+        receiver_center = patches_center[i_receiver]
+        # calculation of form factors
+        receiver_normal = patches_normal[i_receiver]
+        dot_product = np.dot(receiver_normal, source_normal)
+
+        if dot_product == 0:  # orthogonal
+
+            if np.abs(source_normal[0]) > 1e-5:
+                idx_source = set([2, 1])
+                dl = source_center[2]
+                dm = source_center[1]
+                dd_l = patches_size[i_source, 2]
+                dd_m = patches_size[i_source, 1]
+            elif np.abs(source_normal[1]) > 1e-5:
+                idx_source = set([2, 0])
+                dl = source_center[2]
+                dm = source_center[0]
+                dd_l = patches_size[i_source, 2]
+                dd_m = patches_size[i_source, 0]
+            elif np.abs(source_normal[2]) > 1e-5:
+                idx_source = set([0, 1])
+                dl = source_center[1]
+                dm = source_center[0]
+                dd_l = patches_size[i_source, 1]
+                dd_m = patches_size[i_source, 0]
+
+            if np.abs(receiver_normal[0]) > 1e-5:
+                idx_l = 1 if 1 in idx_source else 2
+                idx_s = 0
+                idx_r = 2 if 1 in idx_source else 1
+                dl_prime = receiver_center[1]
+                dn_prime = receiver_center[2]
+            elif np.abs(receiver_normal[1]) > 1e-5:
+                idx_l = 0 if 0 in idx_source else 2
+                idx_s = 1
+                idx_r = 2 if 0 in idx_source else 0
+                dl_prime = receiver_center[0]
+                dn_prime = receiver_center[2]
+            elif np.abs(receiver_normal[2]) > 1e-5:
+                idx_l = 0 if 0 in idx_source else 1
+                idx_s = 2
+                idx_r = 1 if 0 in idx_source else 0
+                dl_prime = receiver_center[1]
+                dn_prime = receiver_center[0]
+
+            dm = np.abs(
+                source_center[idx_s]-receiver_center[idx_s])
+            dl = source_center[idx_l]
+            dl_prime = receiver_center[idx_l]
+            dn_prime = np.abs(
+                source_center[idx_r]-receiver_center[idx_r])
+
+            d = np.sqrt( ( (dl - dl_prime) ** 2 ) + ( dm ** 2 ) + (
+                dn_prime ** 2) )
+
+            # Equation 13
+            A = ( dm - ( 0.5 * dd_m ) ) / ( np.sqrt( ( (
+                dl - dl_prime) ** 2 ) + ( ( dm - (
+                    0.5 * dd_m ) ) ** 2 ) + ( dn_prime ** 2) ) )
+
+            # Equation 14
+            B_num = dm + (0.5 * dd_m)
+            B_denum =  np.sqrt(( np.square(dl - dl_prime) ) + (
+                np.square(dm + (0.5*dd_m)) ) + (
+                    np.square(dn_prime)) )
+            B = B_num/B_denum
+
+            # Equation 15
+            one = np.arctan( np.abs( ( dl - (
+                0.5*dd_l) - dl_prime ) / (dn_prime) ) )
+            two = np.arctan( np.abs( ( dl + (
+                0.5*dd_l) - dl_prime ) / (dn_prime) ) )
+
+            k = -1 if np.abs(dl - dl_prime) < 1e-12 else 1
+
+            theta = np.abs( one - (k*two) )
+
+            # Equation 11
+            ff =  (
+                1 / (2 * np.pi) ) * (np.abs(
+                    (A ** 2) - (B ** 2) )) * theta
+
+        else:
+            # parallel
+            if np.abs(receiver_normal[0]) > 1e-5:
+                dl = receiver_center[1]
+                dm = receiver_center[0]
+                dn = receiver_center[2]
+                dl_prime = source_center[1]
+                dm_prime = source_center[0]
+                dn_prime = source_center[2]
+                dd_l = patches_size[i_source, 1]
+                dd_n = patches_size[i_source, 2]
+            elif np.abs(receiver_normal[1]) > 1e-5:
+                dl = receiver_center[0]
+                dm = receiver_center[1]
+                dn = receiver_center[2]
+                dl_prime = source_center[0]
+                dm_prime = source_center[1]
+                dn_prime = source_center[2]
+                dd_l = patches_size[i_source, 0]
+                dd_n = patches_size[i_source, 2]
+            elif np.abs(receiver_normal[2]) > 1e-5:
+                dl = receiver_center[1]
+                dm = receiver_center[2]
+                dn = receiver_center[0]
+                dl_prime = source_center[1]
+                dm_prime = source_center[2]
+                dn_prime = source_center[0]
+                dd_l = patches_size[i_source, 1]
+                dd_n = patches_size[i_source, 0]
+
+            d = np.sqrt(
+                ( (dl - dl_prime) ** 2 ) +
+                ( (dn - dn_prime) ** 2 ) +
+                ( (dm - dm_prime) ** 2 ) )
+            # Equation 16
+            ff =  ( dd_l * dd_n * ( (
+                dm-dm_prime) ** 2 ) ) / ( np.pi * ( d**4 ) )
+
+        form_factors[i_source, i_receiver] = ff
+    return form_factors
+
+
+@numba.njit(parallel=True)
+def _form_factors_with_directivity(
+        visibility_matrix, form_factors, n_bins, patches_center, air_attenuation,
+        absorption, absorption_index, patch_to_wall_ids,
+        scattering, scattering_index, sources, receivers):
+    """Calculate the form factors with directivity."""
+    n_patches = patches_center.shape[0]
+    form_factors_tilde = np.zeros((n_patches, n_patches, n_patches, n_bins))
+    # loop over previous patches, current and next patch
+
+    for ii in numba.prange(n_patches**3):
+        h = ii % n_patches
+        i = int(ii/n_patches) % n_patches
+        j = int(ii/n_patches**2) % n_patches
+        visible_hi = visibility_matrix[
+            h, i] if h < i else visibility_matrix[i, h]
+        visible_ij = visibility_matrix[
+            i, j] if i < j else visibility_matrix[j, i]
+        if visible_hi and visible_ij:
+            difference_receiver = patches_center[i]-patches_center[j]
+            wall_id_i = int(patch_to_wall_ids[i])
+            difference_receiver /= np.linalg.norm(difference_receiver)
+            ff = form_factors[i, j] if i<j else form_factors[j, i]
+
+            distance = np.linalg.norm(difference_receiver)
+            form_factors_tilde[h, i, j, :] = ff
+            if air_attenuation is not None:
+                form_factors_tilde[h, i, j, :] = form_factors_tilde[
+                    h, i, j, :] * np.exp(-air_attenuation * distance)
+
+            if scattering is not None:
+                scattering_factor = geometry.get_scattering_data(
+                    patches_center[h], patches_center[i], patches_center[j],
+                    sources, receivers, wall_id_i,
+                    scattering, scattering_index)
+                form_factors_tilde[h, i, j, :] = form_factors_tilde[
+                    h, i, j, :] * scattering_factor
+
+            if absorption is not None:
+                source_wall_idx = absorption_index[wall_id_i]
+                form_factors_tilde[h, i, j, :] = form_factors_tilde[
+                    h, i, j, :] * (1-absorption[source_wall_idx])
+
+    return form_factors_tilde
